# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * House Removals Business API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface DeleteResult {
  success: boolean;
}

export type QuotePreferredContact = typeof QuotePreferredContact[keyof typeof QuotePreferredContact];


export const QuotePreferredContact = {
  phone: 'phone',
  email: 'email',
  either: 'either',
} as const;

export type QuoteStatus = typeof QuoteStatus[keyof typeof QuoteStatus];


export const QuoteStatus = {
  pending: 'pending',
  accepted: 'accepted',
  declined: 'declined',
} as const;

export interface Quote {
  id: number;
  name: string;
  phone: string;
  email: string;
  address: string;
  jobDescription: string;
  preferredContact: QuotePreferredContact;
  /** @nullable */
  photoUrl: string | null;
  status: QuoteStatus;
  /** @nullable */
  adminNotes: string | null;
  createdAt: string;
}

export type QuoteInputPreferredContact = typeof QuoteInputPreferredContact[keyof typeof QuoteInputPreferredContact];


export const QuoteInputPreferredContact = {
  phone: 'phone',
  email: 'email',
  either: 'either',
} as const;

export interface QuoteInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  phone: string;
  email: string;
  /** @minLength 1 */
  address: string;
  /** @minLength 1 */
  jobDescription: string;
  preferredContact: QuoteInputPreferredContact;
  photoUrl?: string;
}

export type QuoteUpdateStatus = typeof QuoteUpdateStatus[keyof typeof QuoteUpdateStatus];


export const QuoteUpdateStatus = {
  pending: 'pending',
  accepted: 'accepted',
  declined: 'declined',
} as const;

export interface QuoteUpdate {
  status?: QuoteUpdateStatus;
  adminNotes?: string;
}

export interface QuoteStats {
  total: number;
  pending: number;
  accepted: number;
  declined: number;
  recentCount: number;
}

export interface UploadResult {
  url: string;
}

export interface LoginInput {
  /** @minLength 1 */
  password: string;
}

export interface AuthStatus {
  authenticated: boolean;
}

export interface CalendarEvent {
  id: number;
  /** @nullable */
  quoteId: number | null;
  title: string;
  /** @nullable */
  description: string | null;
  jobDate: string;
  /** @nullable */
  customerName: string | null;
  /** @nullable */
  address: string | null;
  createdAt: string;
}

export interface CalendarEventInput {
  quoteId?: number;
  /** @minLength 1 */
  title: string;
  description?: string;
  jobDate: string;
  customerName?: string;
  address?: string;
}

export type ListQuotesParams = {
status?: ListQuotesStatus;
};

export type ListQuotesStatus = typeof ListQuotesStatus[keyof typeof ListQuotesStatus];


export const ListQuotesStatus = {
  pending: 'pending',
  accepted: 'accepted',
  declined: 'declined',
} as const;