# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: House Removals Business API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: quotes
    description: Quote request operations
  - name: auth
    description: Admin authentication
  - name: calendar
    description: Calendar event operations
  - name: uploads
    description: File upload operations

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /quotes:
    get:
      operationId: listQuotes
      tags: [quotes]
      summary: List all quote requests (admin)
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [pending, accepted, declined]
      responses:
        "200":
          description: List of quote requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Quote"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createQuote
      tags: [quotes]
      summary: Submit a new quote request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuoteInput"
      responses:
        "201":
          description: Quote request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quote"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /quotes/{id}:
    get:
      operationId: getQuote
      tags: [quotes]
      summary: Get a single quote request (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Quote request details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quote"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateQuote
      tags: [quotes]
      summary: Update quote status or notes (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuoteUpdate"
      responses:
        "200":
          description: Updated quote
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quote"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /quotes/stats:
    get:
      operationId: getQuoteStats
      tags: [quotes]
      summary: Get quote statistics (admin dashboard summary)
      responses:
        "200":
          description: Quote statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuoteStats"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      operationId: adminLogin
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthStatus"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: adminLogout
      tags: [auth]
      summary: Admin logout
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthStatus"

  /auth/me:
    get:
      operationId: getAuthStatus
      tags: [auth]
      summary: Check admin session status
      responses:
        "200":
          description: Auth status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthStatus"

  /calendar:
    get:
      operationId: listCalendarEvents
      tags: [calendar]
      summary: List all calendar events (admin)
      responses:
        "200":
          description: List of calendar events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CalendarEvent"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createCalendarEvent
      tags: [calendar]
      summary: Add a job to the calendar (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CalendarEventInput"
      responses:
        "201":
          description: Calendar event created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalendarEvent"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /calendar/{id}:
    delete:
      operationId: deleteCalendarEvent
      tags: [calendar]
      summary: Remove a calendar event (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Event deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    DeleteResult:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success

    Quote:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        jobDescription:
          type: string
        preferredContact:
          type: string
          enum: [phone, email, either]
        photoUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, accepted, declined]
        adminNotes:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - name
        - phone
        - email
        - address
        - jobDescription
        - preferredContact
        - photoUrl
        - status
        - adminNotes
        - createdAt

    QuoteInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        phone:
          type: string
          minLength: 1
        email:
          type: string
        address:
          type: string
          minLength: 1
        jobDescription:
          type: string
          minLength: 1
        preferredContact:
          type: string
          enum: [phone, email, either]
        photoUrl:
          type: string
      required:
        - name
        - phone
        - email
        - address
        - jobDescription
        - preferredContact

    QuoteUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [pending, accepted, declined]
        adminNotes:
          type: string

    QuoteStats:
      type: object
      properties:
        total:
          type: integer
        pending:
          type: integer
        accepted:
          type: integer
        declined:
          type: integer
        recentCount:
          type: integer
      required:
        - total
        - pending
        - accepted
        - declined
        - recentCount

    UploadResult:
      type: object
      properties:
        url:
          type: string
      required:
        - url

    LoginInput:
      type: object
      properties:
        password:
          type: string
          minLength: 1
      required:
        - password

    AuthStatus:
      type: object
      properties:
        authenticated:
          type: boolean
      required:
        - authenticated

    CalendarEvent:
      type: object
      properties:
        id:
          type: integer
        quoteId:
          type: ["integer", "null"]
        title:
          type: string
        description:
          type: ["string", "null"]
        jobDate:
          type: string
        customerName:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - quoteId
        - title
        - description
        - jobDate
        - customerName
        - address
        - createdAt

    CalendarEventInput:
      type: object
      properties:
        quoteId:
          type: integer
        title:
          type: string
          minLength: 1
        description:
          type: string
        jobDate:
          type: string
        customerName:
          type: string
        address:
          type: string
      required:
        - title
        - jobDate