# lib/api-zod/src/generated/types/calendarEvent.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * House Removals Business API
 * OpenAPI spec version: 0.1.0
 */

export interface CalendarEvent {
  id: number;
  /** @nullable */
  quoteId: number | null;
  title: string;
  /** @nullable */
  description: string | null;
  jobDate: string;
  /** @nullable */
  customerName: string | null;
  /** @nullable */
  address: string | null;
  createdAt: string;
}