# lib/api-zod/src/generated/types/calendarEventInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * House Removals Business API
 * OpenAPI spec version: 0.1.0
 */

export interface CalendarEventInput {
  quoteId?: number;
  /** @minLength 1 */
  title: string;
  description?: string;
  jobDate: string;
  customerName?: string;
  address?: string;
}