# lib/api-zod/src/generated/types/loginInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * House Removals Business API
 * OpenAPI spec version: 0.1.0
 */

export interface LoginInput {
  /** @minLength 1 */
  password: string;
}