# lib/api-zod/src/generated/types/quote.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * House Removals Business API
 * OpenAPI spec version: 0.1.0
 */
import type { QuotePreferredContact } from './quotePreferredContact';
import type { QuoteStatus } from './quoteStatus';

export interface Quote {
  id: number;
  name: string;
  phone: string;
  email: string;
  address: string;
  jobDescription: string;
  preferredContact: QuotePreferredContact;
  /** @nullable */
  photoUrl: string | null;
  status: QuoteStatus;
  /** @nullable */
  adminNotes: string | null;
  createdAt: string;
}