# lib/api-zod/src/generated/types/quoteInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * House Removals Business API
 * OpenAPI spec version: 0.1.0
 */
import type { QuoteInputPreferredContact } from './quoteInputPreferredContact';

export interface QuoteInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  phone: string;
  email: string;
  /** @minLength 1 */
  address: string;
  /** @minLength 1 */
  jobDescription: string;
  preferredContact: QuoteInputPreferredContact;
  photoUrl?: string;
}