# lib/api-zod/src/generated/types/quoteStats.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * House Removals Business API
 * OpenAPI spec version: 0.1.0
 */

export interface QuoteStats {
  total: number;
  pending: number;
  accepted: number;
  declined: number;
  recentCount: number;
}