# lib/api-zod/src/generated/types/quoteUpdate.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * House Removals Business API
 * OpenAPI spec version: 0.1.0
 */
import type { QuoteUpdateStatus } from './quoteUpdateStatus';

export interface QuoteUpdate {
  status?: QuoteUpdateStatus;
  adminNotes?: string;
}