# lib/db/src/schema/quotes.ts
import { pgTable, serial, text, timestamp, pgEnum } from "drizzle-orm/pg-core";
import { createInsertSchema } from "drizzle-zod";
import { z } from "zod/v4";

export const quoteStatusEnum = pgEnum("quote_status", ["pending", "accepted", "declined"]);
export const preferredContactEnum = pgEnum("preferred_contact", ["phone", "email", "either"]);

export const quotesTable = pgTable("quotes", {
  id: serial("id").primaryKey(),
  name: text("name").notNull(),
  phone: text("phone").notNull(),
  email: text("email").notNull(),
  address: text("address").notNull(),
  jobDescription: text("job_description").notNull(),
  preferredContact: preferredContactEnum("preferred_contact").notNull().default("either"),
  photoUrl: text("photo_url"),
  status: quoteStatusEnum("status").notNull().default("pending"),
  adminNotes: text("admin_notes"),
  createdAt: timestamp("created_at").notNull().defaultNow(),
});

export const insertQuoteSchema = createInsertSchema(quotesTable).omit({ id: true, status: true, adminNotes: true, createdAt: true });

export type InsertQuote = z.infer<typeof insertQuoteSchema>;
export type Quote = typeof quotesTable.$inferSelect;